Restructure code to allow tests to imports that then import further
